question list and delete

+++ field-audit-web/src/main/java/com/jianla/faudit/controller/QuestionController.java
@@ -4,10 +4,12 @@ import com.jianla.faudit.dto.QuestionDto;
 import com.jianla.faudit.entity.Question;
 import com.jianla.faudit.model.QuestionForm;
 import com.jianla.faudit.service.QuestionService;
+import com.jianla.faudit.service.ServiceQuestionService;
 import com.jianla.faudit.util.AuthUtil;
 import com.jianla.model.base.JsonResult;
 import com.jianla.model.base.Page;
 import com.jianla.model.base.SessionInfo;
+import com.jianla.util.ParseUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +18,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
+
+import java.util.List;
 
 /**
  * @author : zwj
@@ -29,6 +33,8 @@ public class QuestionController {
 
     @Autowired
     private QuestionService questionService;
+    @Autowired
+    private ServiceQuestionService serviceQuestionService;
 
     @RequestMapping(value = "",method = RequestMethod.GET)
     public String listPage(){
@@ -50,7 +56,7 @@ public class QuestionController {
      */
     @RequestMapping(value = "data",method=RequestMethod.GET)
     @ResponseBody
-    public JsonResult list(SessionInfo sessionInfo, Page page,QuestionDto question){
+    public JsonResult list(SessionInfo sessionInfo, Page<QuestionDto> page,QuestionDto question){
         question.setOrgId(sessionInfo.getOrgnizationId());
         page = questionService.find(page, question);
         JsonResult result = new JsonResult();
@@ -106,12 +112,22 @@ public class QuestionController {
      */
     @RequestMapping(value = "{id}/delete",method = RequestMethod.POST)
     @ResponseBody
-    public JsonResult doDelete(@PathVariable("id") Long id){
+    public JsonResult doDelete(@PathVariable("id") Long id,SessionInfo sessionInfo){
+
         try {
+            Question question = questionService.getById(id);
+            AuthUtil.assertEqual(sessionInfo.getOrgnizationId(),question.getOrgId());
+
+            List<Long> serviceIds = serviceQuestionService.findServiceIdByQuestionId(id);
+            if(serviceIds.size()!=0){
+                return new JsonResult(false,"无法删除问题，下列服务引用【"+ ParseUtil.toString(serviceIds)+"】");
+            }
+
             questionService.delete(id);
             return JsonResult.OK;
         } catch (Exception e) {
-            return new JsonResult(false,e.getMessage());
+            logger.error("删除问题出错",e);
+            return new JsonResult(false,"删除问题出错");
         }
     }
